refactoring methods into traits
